+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.8.0
 * UI improvement: Better "CSVLint" status button by [@kostasx](https://github.com/kostasx).
+* Update RBQL: Add `ANY_VALUE` aggregate function.
 
 ## 3.7.1
 * Update RBQL: Fix GROUP BY queries with bare columns, see [#144](https://github.com/mechatroner/vscode_rainbow_csv/issues/144)